Grids, errorbars and scale
Added a grid for each bin in x, tidied up the plots increasing their axis and title font sizes. Added errorbars to the plot and added lines to change scales to logarithmic, however in my opinion this looks worse than a linear scale. Also added an argument 'close' to the function makeplot that closes graphs that open in Spyder.